Replace .split(.) with os.path.splitext() in checkfile function.
In the checkfile function, the file extension is extracted from the file_location string using the split() method. This method can be error-prone because it only splits the string at the first occurrence of the specified delimiter. Therefore, it may not work correctly if the file location contains multiple dots.

To avoid this issue, we replaced the split() method with the os.path.splitext() function, which splits the path into a tuple of the file's base name and extension. This ensures that we always get the correct file extension regardless of the number of dots in the file name.